# sesi-08/swaggo/docs/swagger.yaml
basePath: /
definitions:
  main.Item:
    properties:
      description:
        example: A random description
        type: string
      itemId:
        example: A1B2C3
        type: string
      quantity:
        example: 1
        type: integer
    type: object
  main.Order:
    properties:
      customerName:
        example: Leo Messi
        type: string
      items:
        items:
          $ref: '#/definitions/main.Item'
        type: array
      orderId:
        example: "1"
        type: string
      orderedAt:
        example: "2019-11-09T21:21:46+00:00"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
  description: This is a sample service for managing orders
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Orders API
  version: "1.0"
paths:
  /orders:
    get:
      consumes:
      - application/json
      description: Get details of all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Order'
            type: array
      summary: Get details of all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with the input paylod
      parameters:
      - description: Create order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/main.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Order'
      summary: Create a new order
      tags:
      - orders
  /orders/{orderId}:
    delete:
      consumes:
      - application/json
      description: Delete the order corresponding to the input orderId
      parameters:
      - description: ID of the order to be deleted
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete order identified by the given orderId
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Get details of order corresponding to the input orderId
      parameters:
      - description: ID of the order
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Order'
      summary: Get details for a given orderId
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update the order corresponding to the input orderId
      parameters:
      - description: ID of the order to be updated
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Order'
      summary: Update order identified by the given orderId
      tags:
      - orders
swagger: "2.0"